📦 chore : 이미지 API 인스턴스 분리 -> fastClient
