+++ src/apis/client.ts
@@ -11,14 +11,6 @@ export const cleanHeaderInstance = axios.create({
   headers: {},
 });
 
-// [이미지 업로드]
-// 8002 포트번호 다른 이유로 인스턴스 생성
-export const imageinstance = axios.create({
-  baseURL: import.meta.env.VITE_APP_IMG_URL,
-  withCredentials: false,
-  headers: {},
-});
-
 export const instance = axios.create({
   baseURL: import.meta.env.VITE_APP_BASE_URL,
   withCredentials: true,
@@ -34,7 +26,7 @@ export function post<T>(...args: Parameters<typeof instance.post>) {
 }
 
 export function put<T>(...args: Parameters<typeof instance.put>) {
-  return imageinstance.put<T>(...args);
+  return instance.put<T>(...args);
 }
 
 export function patch<T>(...args: Parameters<typeof instance.patch>) {
@@ -44,7 +36,3 @@ export function patch<T>(...args: Parameters<typeof instance.patch>) {
 export function del<T>(...args: Parameters<typeof instance.delete>) {
   return instance.delete<T>(...args);
 }
-
-// export function put<T>(...args: Parameters<typeof instance.put>) {
-//   return instance.put<T>(...args);
-// }
