+++ src/components/PostNew/Step6/Step6.tsx
@@ -5,9 +5,9 @@ import food1 from '../../../assets/Image/food1.jpg';
 import food2 from '../../../assets/Image/food2.jpg';
 import food3 from '../../../assets/Image/food3.jpg';
 import { IcEmptyThumbnailFinal, TipBtn } from '../../../assets/svg';
-import { put } from '../../../apis/client';
 import { useOnboardingContext } from '../../../context/PostNew/PostNewContext';
 import usePostOnboardingInfo from '../../../queries/PostNew/usePostInfo';
+import { put } from '../../../apis/fastClient';
 
 interface ServerResponse {
   file_name: string;
